Copy and modify ownership of ssl key file
Certificate key file has to be accessible by icecast user. The one on the machine is usually owned by root, so icecast cannot read it. We copy it to the effemeral storage and chown it, then modify icecast.xml to use it.

At the same time we copy icecast.xml to icecast-owned.xml on the effemeral storage, so the modifications to apply the environmental variables are not written to the original file.